Adds `site` example with basic home page.
This will act as a more complete example than most of the others and provide an opportunity to the use the toolchain for a real-ish site to see how well it works.

First thing I'm discovering is that each page needs its own `web_resources_devserver()` in order to be tested by Puppeteer. That's only if users want to add end-to-end tests for pages, which we do in `rules_prerender` because it is critical to our functionality, but not every site will feel the need to do it. Unit tests of JavaScript and maybe some screenshot tests may be enough for most sites.