Adds `base` layout.
This provides the HTML structure for a simple page, reducing boilerplate repeated across different pages.

It mostly works as well as expected, although the one interesting problem I had not considered is user testing of individual components. These are too small and independent to use a Puppeteer test for. However, I think I found a decent middle ground in simply using an HTML parser with `document.querySelector()` to identify specific elements and assert on them.

This test strategy works well enough, although the parser library is a little awkward to work with. The `<main />` element for instance actually has three children, two of which are text nodes with just whitespace and requiring extra code to ignore.

There's also an argument to be made that this module is so simple as to not require tests in the first place. All I'm doing is asserting a basic HTML structure which is quite likely to change, meaning the test is somewhat brittle. Most people don't write tests for their HTML (or if they do, it's more about an end-to-end integration test or screenshot test). I'm personally of the opinion that if you wrote a piece of code you should have some automated piece of infrastructure that tells you it works as expected. Templates like this are code and should be tested similarly IMHO. Of course it is up to application developers to decide how much effort they would want to put in to testing components. For something like this, I can absolutely understand not bothering with writing in the test, and maybe only doing so for components with complex logic (like generating pages from Markdown, or something like that).

The one broken abstraction here is that I need a direct dependency on `%{component_name}_prerender`. I've hestiated considering this library public API because you should depend on all of its ancillary resources as well, while depending on just the prerender library won't do that. For testing templates like this, it's not too big a deal because it won't be loaded into a browser anyways. We may want to export that, at least for testing purposes if necessary.