Adds a header component with basic styling (it's ugly I know, don't judge).
The header component itself is relative straightforward, just a site title with some navigation items. Currently there is only a home page, so that's the only one present.

Most of the work here is in styling. I've added CSS files and marked the top-level container element of each component with a `comp-*` attribute. This is used by styling to limit their effects to only the one component. Until declarative shadow DOM is properly spec'd out, there's not much better we can do here.

I opted for an attribute rather than a class name so it could be referenced from JavaScript without worrying about tooling renaming classes. It also includes a `comp-` prefix to avoid conflicts with other, unrelated attributes.

Some global constants are kept in a `theme.css` file, which is a simple filegroup depended upon by each component that imports it. The import works pretty well and nicely abstracts the content in an easily bundled fashion, however the lack of strict deps makes this quite fiddly. If not for writing this part of the commit message, I would have forgotten to make `home.css` dependent on `theme.css`. I'm hoping Sass infrastructure can provide a bit more tooling here.