Adds an "About" page.
Just placeholder content for now, the main purpose here is just to introduce an additinal page and add a link to the header.

I'm deliberately **not** adding tests for this page because I think the best way to test a page like this is through screenshots, which I don't have set up atm and don't want to deal with. Tests for pages like this will come later.