Moves `packages/rules_prerender/index.bzl` to `index.bzl`.
Motivation for this change is to support publishing to NPM. The `pkg_npm()` rule sets up directories in the same pattern as the source tree relative to itself. This means that `pkg_npm()` needs to be in the root directory in order to access all the relevant tools and keep them in the right locations. Based on that, `index.bzl` needs to be at the root of the repository in order to be importable at `@npm//rules_prerender:index.bzl`.