Adds `//tools/internal` package.
This package serves as a singular location where tools needed at build time for this repository and at build time in user repositories can be colocated with different implementations. The `BUILD.bazel` file provides the implementation for `bazel` commands for contributors of `rules_prerender`. The `BUILD.publish` file will be published to NPM with the built package, thus it updates paths to use `@npm//rules_prerender` and makes the same tools work in that environment. This allows Bazel rules and macros to refer to tools at a singular location `//tools/internal:*` and trust that the tool will be available for local contributors and also for users of the release build.

Currently `BUILD.publish` isn't used anywhere, but it will be included in the `pkg_npm()` target that exports the `rules_prerender` package.

This also updates VSCode settings to treat `BUILD.publish` as a Starlark file, so it should get syntax highlight and intellisense on par with `BUILD.bazel` files.