Adds `pkg_npm()` target.
This packages the repository up for deployment to NPM. It sets up all the relevant packages in the repository with `publish_files()` rules to clone all the required files and tools for runtime execution as well as the BUILD files to coordinate them.

Most of the tools are depended upon directly by `pkg_npm()`, which drops all their transitive source files directly in the package. Other tools like `*.bzl` files, the Rollup config, and PostCSS plugin configuration need to be manually included and exported at runtime.

The `BUILD.publish` files are automatically included by `publish_files()` and inserted into the NPM package to be used at runtime by user code. When a user loads `@npm//rules_prerender/...`, they will be seeing the `BUILD.publish` files, while a contributor to `rules_prerender` will use `BUILD.bazel` files.

More complete documentation is coming in a future commit. This was tested with a separate Bazel workspace that `npm install`-ed the `*.tgz` file output by `bazel run //:pkg.pack`, which is the easiest way to try it out from the user perspective. Long term, I have lofty goals of how to test this in an automated fashion, but for now this will need to be done manually as proper automation will likely be quite involved to build. For now we just have a simple `build_test()`.