Drops `type="module"` from `script`.
I forgot that older browsers simply won't execute script tags with `type="module"` and I don't want to both with differential loading at the moment.

Most of the benefits of modules are minor and not entirely applicable here. We bundle everything, so no `import` or `export` statements will make it to the browser. The site is very likely to be hosted on the same origin, so CORS isn't an issue, we are still `defer`-ing by default, not repeating script tags, or inlining them. As a result, there is very little benefit to using `type="module"` but it causes a small, but non-trivial hit to browser support.

Reference: https://jakearchibald.com/2017/es-modules-in-browsers/