Adds release configuration which optimizes and stamps builds with a NPM version number.
See: https://bazelbuild.github.io/rules_nodejs/stamping.html

This provides a `--config release` flag for Bazel which will optimize and stamp builds. The stamping process also accepts an NPM semver as input (via the `$BAZEL_RULES_PRERENDER_VERSION` environment variable) and applies it to the `package.json` in the published NPM package.

This goes a bit against Bazel-ideology, as it is impossible to stamp the built NPM package without an external environment variable, however this allows us to build a usable package with the version as an input to the build. This means we don't need to commit a change to update the version number every time.

This should be usable from a manually invoked GitHub action, which should allow us to explicitly specify the version number of the given release.

The build process is a little unique, as it seems that the `substitutions` attribute doesn't have useful `--nostamp` behavior. I filed an issue about this, but for now we can use the best workaround I could identify, which is unnecessarily complicated but essentially involves using different substitutions for stamped and unstamped builds.

Having now written this commit and solved all the problems associated with it, I think there might be a better way. I think we could hard code the major and minor version numbers and then generate the patch version based on the Git branch. Git commits are SHA1 hashes and are not monotonic, so they are not directly usable, but we could do something like `git log --oneline | wc -l` to assign a revision number of the current branch. This could be used as the patch value (or subtract a hard-coded initial value to have smaller patch values). Such a strategy could work better when managing multiple branches, each of which would have different major and minor version values, but would compute their own patch number in a monotonic compatible way. It would also mean that releases can be entirely automated and no human intervention would be required to choose a version number.

That said, we're not at that level of complexity yet, so I'll stick with this strategy for now. We can re-evaluate and update to that strategy in the future if/when it becomes helpful.