Adds `--bazel_patch_module_resolver` to all NodeJS binaries and tests.
In the upcoming `rules_nodejs` v3 release, this flag will be disabled by default, however it is still necessary to import modules in NodeJS. The easiest solution is to just explicitly enable the flag, which is a no-op in v2, but will prevent the breaking change in v3.

Unfortunately I have no idea how to actually fix the underlying breakage, only work around it. I found [an issue](https://github.com/bazelbuild/rules_nodejs/issues/2388) which seems to represent this particular problem, will keep an eye on it and update this later with a real fix that does not require this flag.