Upgrades workspace to use `rules_nodejs` v3.
[Documentation of changes.](https://github.com/bazelbuild/rules_nodejs/releases/tag/3.0.0)

The upgrade is mostly done by just updating the `rules_nodejs` `http_archive()` rule to use the `3.0.0` version and updating `@bazel/*` NPM dependencies to be `^3.0.0`.

The biggest breaking change that could not be addressed in an earlier commit is that `ts_devserver()` was renamed to `concatjs_devserver()` and moved to `@bazel/concatjs`. The rename isn't that difficult, but `@bazel/karma` infrastructure got moved to the same package, meaning we now need a dependency on `@io_bazel_rules_webtesting` to satisfy the Karma rules in `@bazel/concatjs`, even though we don't use them. There are also a number of peer dep issues, as we're supposed to install Karma-related packages. I'm just ignoring those warnings for now as we definitely don't use Karma atm.

Another side effect of this is that `rules_webtesting` is required for `@bazel/concatjs` to load. For this workspace that's not too big a deal, but it is also necessary for user workspaces since `web_resources_devserver()` now uses `concatjs_devserver()` from `@bazel/concatjs`, which itself requires `rules_webtesting`. This unfortunately means all user projects have an awkward dependency on `rules_webtesting`. On the bright side, it seems we don't need to actually call a function from it (per [`rules_nodejs@3` release notes](https://github.com/bazelbuild/rules_nodejs/releases/tag/3.0.0)), so we can add this workspace in `rules_prerender_dependencies()`.

Added a call to `rules_prerender_dependencies()` to the local `WORKSPACE` file which probably should have always been there. This calls through to NPM deps like `rules_postcss` to install transitive deps.

We also had to set `strict_visibility = False` so `rules_postcss` can resolve its NPM dependencies. I don't fully understand the problem here (as I think it is an issue with `rules_postcss` rather than anything I'm doing wrong), but my guess is that `rules_postcss` exports macros like `postcss_binary()` which generates a `nodejs_binary()` rule that depends on PostCSS deps (such as `@npm//minimist`). PostCSS does declare that dependency appropriately, but because the macro gets executed and evaluated in the `rules_prerender` workspace, which does not have a dependency on `minimist`, strict visiblity prevents the dep and fails the build. Unfortunately, I'm not sure how exactly to fix this at the moment, so the easiest solution is to just disable strict visibility.