Adds tests for `label.bzl`.
I forgot that Starlark utility files can be pretty easily tested and didn't include tests when it was first written. Since this just transforms strings, it's pretty easily tested.

Only hiccup was that `native.package_name()` and `native.repository_name()` can't be called at analysis time, which is how the tests are executed. Modified the function to accept those as optional inputs so make this testable. Unfortunately these need to be functions, not raw strings, which means that we need top-level named functions to mock these values in tests because Starlark does not support `lambda` syntax nor inner functions.

https://docs.bazel.build/versions/master/skylark/testing.html#for-testing-starlark-utilities