Adds `file_path_of` utility.
This converts a target label to a simple path. This path would represent the runtime path to that target, assuming the target generates a file or directory at its `name` attribute.

This is useful to better test edge cases around external workspaces and root packages, which require special logic to omit slashes that might otherwise be included.

Motivation for this is that `web_resources_devserver()` has a bug that its path does not work correctly in the root BUILD file, due to that edge case not being totally covered. It will be updated to use this utility which has better test coverage for these cases.

I wanted to author this such that it will always emit the workspace name as the first directory, however I'm not aware of any way to actually get the workspace name from the context of a macro. `native.repository_name()` and `Label().workspace_name` both return `"@"` for the main workspace, and `__main__` does not work either. Instead, if called in the main workspace, it will just not include the workspace name. That works for `concatjs_devserver()`, so it will hopefully be good enough for now.