Updates README to be more clear about installation and usage for v0 release.
Opted to drop the peer deps suggestion because it mostly just complains about Karma, which doesn't matter, and also that `@bazel/typescript` must have `typescript` < 4.0, which is a weird setup that we can't really avoid.

Added a suggestion to make a `tsconfig.json` since that is the first error most users would hit.

Beyond that, most of this is general cleanup. Getting started docs are still pretty minimal, but for a v0 release that is somewhat expected.