Adds example which verifies custom `tsconfig` files.
This simply sets `tsconfig` on `prerender_page_bundled()` (which composes `prerender_page()` and `prerender_component()` under the hood). This custom `tsconfig` is effectively the same as what we have at the repository root, expect it has `strict: false`. We then add a dummy parameter to the default function which is allowed only because `--noImplicitAny` is not set, due to the lack of `--strict`. A `build_test()` rule simply verifies that this builds which is good enough to confirm that `tsconfig` is working as expected.