Adds publish GitHub action.
Refs #14.

This is an initial implementation of a GitHub action which can be manually invoked to publish the package. It accepts a semver as an input and then tests the repository before publishing to NPM.

Like all GitHub actions, it's very hard to test before commiting, so I have no idea if this will actually work.