Removes `body` input from publish GitHub action.
As of today, GitHub actions only supports string inputs. We handle prerelease by simply comparing with a `"true"` string literal, but the body input is impractical to use a single line. The easier solution is to omit user input here and let users edit the release afterwards in the normal GitHub UI.