Rename NPM script from `publish` to `release`.
Refs #14.

Think I figured out the problem. I was very confused why the working directory would be a problem for Bazel and why the script appeared to publish successfully despite failing. It seems that running `npm run publish` works as expected, executing `bazel run //:pkg.publish` which itself calls `npm publish`. Part of the `npm publish` lifecycle is to invoke the user `publish` script. This means that `npm publish` itself ran `npm run publish`, restarting the process and invoking `bazel run //:pkg.publish` again. The only major difference is that it was in the `bazel-out/` directory where the packed package was, which Bazel didn't like and would fail. That also explains why running `bazel run //:pkg.publish` directly in GitHub actions didn't fix anything, because `npm publish` was still calling the `npm run publish` script which I thought wasn't used.

https://docs.npmjs.com/cli/v6/using-npm/scripts#npm-publish

The easiest solution here is to just rename the script to something which does not have special semantics from `npm`. Renaming it to `release` which seems to fit this criteria, even if it doesn't fully match the general terminology I've been trying to stick to.