Removes debugging pieces of publish scripts.
Refs #14.

Last run was successful and it seems `npm publish` was invoking `npm run publish` in an unintended fashion. There should be no more need of all the debugging output and it should be removable.

Also to point out, as part of the process I previously added `--` to the end of the `npm run release` script. This means that any subsequent flags provided by a user (such as `--tag next`) which be passed to the `npm publish` command that gets executed, rather than the `bazel` command's build flags.