Adds `PrerenderResource`.
Refs #1.

This type holds file's contents and its associated URL path. This will be leveraged by users to associate a particular path to its file contents, which `rules_prerender` tooling will generate. It is exported as public API because users will construct the object to pass to `rules_prerender` to generate it.

The file contents are expressed as an `ArrayBuffer` to support binary files without requiring a specific representation like `Uint8Array` would. `string` construction is supported, just bound to UTF-8 encoding, which is usually what users want. If users care about a specific encoding, they can easily convert a string to their own `ArrayBuffer` using their desired encoding and just pass that in, so I don't think binding `string` inputs to UTF-8 will be a maintenance issue.

I opted for `ArrayBuffer` as the type representing binary files because `Blob` doesn't appear to be supported by NodeJS and I wanted to use something that is actually a part of the web spec. `Buffer` is the obvious NodeJS choice, but I didn't see much value in binding to the NodeJS API when an open web specification alternative was just as usable. The downside of not using `Blob` is that file contents are forced to be in-memory and are not easily streamable. I think this is fine for now, but in the future we may want to switch to `Stream`, which would couple us to NodeJS, or using an `AsyncIterable<string | ArrayBuffer>`, which would be a bit less common but at least use standardized APIs.

The most awkward part here is that `Uint8Array` is structurally compatible with `ArrayBuffer` but is **not** a subclass of it. This means that the following compiles and runs without error:

```typescript
function normalize(input: ArrayBuffer | string): ArrayBuffer {
    if (input instanceof ArrayBuffer) return input;
    return new TextEncoder().encode(input);
}

console.log(normalize(new Uint8Array([ 0, 1, 2, 3 ])));
// Uint8Array(7) [ 48, 44, 49, 44, 50, 44, 51 ] - Wat?
```

Because `Uint8Array` is not a subclass of `ArrayBuffer`, the `instanceof` check returns `false`. `TextEncoder` happily encodes the value, effectively encoding the input as a UTF-8 string, when it is intended to be left as is. This is particularly rough because the input comes from user code, so users are very likely to pass in a `Uint8Array` and TypeScript won't complain. The best solution to this problem is to just accept all `TypedArray` types as an input, even though `rules_prerender` truly doesn't care about any of them and we only convert to an `ArrayBuffer` anyways. This unfortunately leaks into our public API and means we have to support another input type, but the alternative is a bunch of wasted debugging time for users which I think is the worse option anyways.