Adds multi-renderer tool.
Refs #1.

This tool is roughly equivalent to the existing renderer tool in purpose, however it allows user code to return `PrerenderResource` objects and will write all of them to disk under a single output directory. This effectively allows users to render multiple files at once rather than being limited to just one.

This tool is exported to users, though it has the same awkward caveats as the existing renderer. Because we dynamically import user code, we must compile user code into the `nodejs_binary()` which runs the tool. As a result, we can't publish a `nodejs_binary()` directly, but must instead build one as needed, including both multi-renderer code and user code all in a single binary. This means the published version of the tool is not actually a `nodejs_binary()` as one might expect, but instead it is just a `filegroup()` of the code required to run the tool. For simplicity we "round up" to all the code in the package, since it makes little difference to users who have to install the whole package anyways and ensures we won't forget to list a critical file.