Adds `prerender_multi_page()` macro.
Refs #1.

This is mostly a copy of `prerender_page()`, except outputting a directory of generated files rather than just a single file. Currently, we don't have the infrastructure to generate entry points for scripts or styles within the context of multiple generated pages. For now, `prerender_multi_page()` just emits an aggregated set of scripts and styles with no specific entry point. Support for that will come later.