Adds multi-annotation extractor tool.
Refs #1.

This tool acts as a new version of the existing annotation extractor tool which supports processing multiple input files at once. It accepts a directory of input files and extracts annotations from all of them, combining them into a single metadata JSON file. The input files are copied to an output directory with annotations stripped out. Non-HTML files are copied directly over with no changes.

Ideally, we should have a unique metadata JSON file *for each* input HTML file. However, `rollup_bundle()` and `postcss_binary()` do not have usable implementations which support bundling multiple files when the file set is not known until execution time. As a result, having unique metadata files is not useful at the moment, so instead we generate one metadata file that contains all the scripts and styles from all HTML files in the input. This is overly broad since if one HTML file includes a script, that script is attributed to *every* HTML file. We will explore making this more preceise and enabling better tree-shaking of unused scripts and styles at a later time.