Moves deduplication from `extractor.ts` to `annotation_extractor.ts`.
Refs #1.

We are going to add a new multi-annotation extractor binary which also needs to deduplicate annotations. However the multi-annotation extractor must make multiple calls to `extract()` for each file it is processing. This means `extract()`'s deduplication is not sufficient, because the file set of all extracted annotations across all processed files must be deduplicated. As a result, `extractor.ts`'s deduplicate logic is not actually useful. Instead this needs to be done at the binary level. This commit simply moves the logic from `extractor.ts` to `annotation_extractor.ts`. A similar implementation will be included in the new multi-annotation extractor.

I also took the opportunity to update the implementation to use the new common `unique()` function with a real `annotationEquals()` implementation rather than relying on Node's deep equality assertion. This is cross-platform and a more authoritative implementation.