Adds `prerender_multi_page_bundled()`.
Refs #1.

This introduces a new `prerender_multi_page_bundled()` as a corollary to `prerender_page_bundled()`, generating multiple files. It calls through to `prerender_multi_page()` to generate the HTML, JS, and CSS files, using Rollup and PostCSS to bundle these resources into individual JS and CSS files. The resource injector tool is currently not compatible with injecting multiple files, so that step is skipped for now. Instead the JS and CSS are bundled, but then left alone and not used. Only the HTML and resources are merged together.

Updated the multi-page example to use `prerender_multi_page_bundled()`. Functionally, all this means for now is that we can drop the excess `web_resources()` target. Also added a `build_test()` for the bundled JS and CSS since nothing else uses them at the moment.

I also took the opportunity to update the documentation of the pre-existing `prerender_multi_page()`, `prerender_page()`, and `prerender_page_bundled()` macros to be consistent between them and fix any minor copy-paste errors I found.