Updates multi-resource injector to actually support injecting resources into multiple files.
Refs #1.

This modifies the existing logic to accept an input directory and output directory as arguments. For every file in the input directory, it is injected with the provided resources and then written to the same relative path in the output directory. Non-HTML files are copied to the output directory unchanged. We unfortunately need to support non-HTML files because a user might generate other files in the same tool which generates HTML, so we can't assume the input directory only has HTML files.

Copy-pasted some helper utilites from multi-annotation extractor for iterating over input files and creating the parent directory for output files. If these utilites continue to be useful in other contexts, it may be worth factoring them out to a shared library.