Adds `multi_inject_resources()` macro and calls it from `prerender_multi_page_bundled()`.
Refs #1.

`multi_inject_resources()` is a rough copy of `inject_resources()` except it uses directories for inputs and outputs. It returns a `WebResourceInfo()` in order to be compatible with `web_resources()`. The multi-resource injector tool is also published to the final NPM package directory. Tested this with the multi-page example. This requires that all HTML files *actually* adhere to the HTML spec, so I had to update the example to create properly formatted HTML files. Also tested with `ref/external` and confirmed that `multi_inject_resources()` and the multi-resource injector tool both function when used by external repositories via `@npm//rules_prerender`.

Currently, `multi_inject_resources()` only deals with CSS files, as this is all that is supported by the multi-resource injector tool. The rule will be updated to support scripts once the tool is able to inject JavaScript in a usable manner.