Adds `bundle` option to multi-resource injector tool.
Refs #1.

This option gives a single JavaScript file which gets copied alongside every HTML file processed. This allows each file to own and link to its own JS, even if they happen to all be the same. Long term, these files will be different, so while we could optimize things to use a single JS file, it would be temporary anyways.

For now, the JavaScript file is only copied alongside each HTML file, a `<script />` tag is not yet injected into the HTML.