Adds bundle support to `multi_inject_resources()` and `prerender_multi_page_bundled()`.
Refs #1.

This means that `prerender_multi_page_bundled()` will bundle all the associated JavaScript to a single file and then inject it into each HTML page, apply the scripts to all of them.

Updated the existing multi-page example to include visible JavaScript on the page and assert on it in tests.