Adds `data` attribute to relevant `prerender_*()` rules.
Refs #24.

This allows prerender binaries to depend on files from source control and use them as part of the generation process. This is implemented just by passing the values through to the underlying `ts_library()` implementation. This may pose a problem when migrating to `ts_project()`, which doesn't seem to support `data`, but I'll worry about that when we get there. It should be easy to work around by just adding a `filegroup()` if necessary.