Adds example which uses the `data` attribute.
Refs #24.

This stores several different posts in text files within the workspace. This example depends on those files via the `data` attribute and then reads them during rendering to render a new HTML page for each text file. It also generates an index page which links to all the post pages.