Migrates `prerender_page_bundled()` usages to `prerender_multi_page_bundled()`.
Refs #25.

`prerender_page_bundled()` is almost exactly equivalent to `prerender_mulit_page_bundled()`. The only real difference is that the former allows user code to return a string and then uses the `path` attribute to write the file to that location. This is unnecessary, since users can just emit a single file with the path stored there.