Adds `golden_test()`.
This is a simple utility for testing a generated file matches the given contents. The implementation is a straightforward and lightweight `diff` command.

Unfortunately I ran into some really unintuitive Bazel behavior. Apparently arguments to `sh_binary()` and `sh_test()` are not quoted. So a single argument with the value "foo bar" is interpreted by the script as two different arguments "foo" and "bar". This means I had to manually quote the string, and I didn't want to implement Bash escaping logic in Starlark. Instead I just appended quotes and added an assertion that inputs cannot contain a double quote. It's a bit awkward, but it's the simplest solution for now.

See https://github.com/bazelbuild/bazel/issues/13072.