Adds `extract_single_resource()`.
Refs #25.

This rule serves an adapter to replace the main use case of `prerender_page()`. My logic is that `prerender_page()` is equivalent to `prerender_multi_page()` except that it allows a user a direct file label on the output, making subsequent processing a bit easier.

Since this is a uncommon use case, I figured an adapter can work better than a completely different `prerender_*()` rule. Now, users should use `prerender_multi_page()` and generate a single file with any path. Then they can use `extract_single_resource()` to easily copy that file into a known label. All subsequent build commands can depend on that label and become much simpler as a result. The `extract_single_resources()` docstring gives an example of this.

With this adapter, we no longer need to support `prerender_page()`, which enables us to drop a lot of non-multi tool infrastructure, hooray!

Tested this with a new `examples/extract/` directory and used the new `golden_test()` to confirm they files were generated as expected. Unfortunately there is no easy way to test the two error cases (when the resources input has no files, or more than one file).

Arguably there is a valid use case for a generic `extract_resource()` rule, which would copy out a resource at a particular path. This may be useful in the future, however as of now, there is no good way of removing a file from a `web_resources()` rule. So even if you extracted a file at a particular path, and then transform it, there's no way to replace the old one with the new content. In the interest of minimizing scope, I'll leave this out for now, however it may be worth considering in the future. Arguably we could drop `extract_single_resource()` and treat it as a subcase of `extract_resource()`.