Removes `prerender_page()`.
Refs #25.

Users are recommended to use `prerender_multi_page()` and simply render one file. If necessary, they can use `extract_single_resource()` to pull the prerendered file out of its generated directory and get a Bazel label that directly refers to the file. This is useful to pass the file into other tools which may not expect a directory as an input.