Refactors `inject_resources()` to be a subcase of `multi_inject_resources()`.
Refs #25.

This updates `multi_inject_resources()` to support a `scripts` argument which works just like `inject_resources()`. Fortunately this was already supported by the multi-resource injector tool because it reused the implementation originally written for the first resource injector tool. `inject_resources()` now just wraps its single file input into a directory, invokes `multi_inject_resources()`, and then extracts the processed file back out. This presents a single-file in, single-file out API, which is much easier to work with in certain circumstances.

Currently `multi_inject_resources()` is not public API because injecting JavaScript is still very weird due to #26. Once we have that figured out, we can polish the API a little more and give something with a better abstraction. Once that happens, we should revisit `inject_resources()` to make sure it still makes sense. The main motivation for this change now is just to remove the only remaning usage on the original resource injector tool, enabling us to delete it.