Converts examples and documentation to prefer `Generator` over `Iterable`.
After some testing I have decided a best practice is to declare renderer functions to return `Generator<PrerenderResource, void, void>` or `AsyncGenerator<PrerenderResource, void, void>`. This commit updates the examples and documentation to use this pattern consistently.

The reasoning for this is because `Generator` is a subtype of `Iterable`, so it still matches and is valid, however it is more strict with regard to generator functions. `Generator<PrerenderResource>` is roughly equivalent to `Iterable<PrerenderResource>` and in both cases, `return PrerenderResource.of(/* ... */);` and `const foo = yield /* ... */;` will both compile successfully, but will not work at runtime. The renderer does not look at the final return value and does not emit values back to user code on each iteration. As a result, `Generator<PrerenderResource, void, void>` is a more accurate reflection of the runtime behavior. TypeScript is also smart enough to emit a compile-time error if those values are used.

Because this type is more accurate and surfaces more possible mistakes to the user, I recommend always using `Generator` and `AsyncGenerator` with the extra `void` type parameters and consider it a "best practice".