Renames `prerender_multi_page()` -> `prerender_pages_unbundled()`.
This drops the `multi` in favor of a plural `pages` because there is no longer a rule for prerendering a single page. It also adds `unbundled` to disincentivize usage of this rule as compared to `prerender_pages()`. Most users will want automatically bundled pages, so that should be the default. Users who don't are likely to wrap it in their own macros anyways, so including `_unbundled` in the name is not that big a deal.