Adds `_prerender_for_test` to a number of `rules_prerender` macros.
Fixes #10.

This is a public API which provides a reference to the `ts_library()` target which compiles the user's `srcs`. It allows users to easily invoke their prerender code for testing purposes. This is specifically marked `testonly` because it is focused on the single testing use case.

This target's existence technically breaks an invariant of `prerender_component()`, in that it only depends on a single piece of a complex component (it does not include client side scripts, CSS, or othe resources). As a result, I'm trying to restrict and discourage usage like this as much as possible. Unit testing is just the one exception where this really gets in the way, so hopefully this `testonly` target will satisfy that use case sufficiently without providing an abusable hook for users to exploit and workaround the guarantees provided by `prerender_component()`.